+++ packages/playground/react/vite.config.js
@@ -1,6 +1,6 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
-import Inspector from 'vite-plugin-component-inspector'
+import Inspector from 'vite-plugin-dev-inspector'
 
 // https://vitejs.dev/config/
 export default defineConfig({